issue #600: K4 export options in UI added
- Moved Actions into separate folder
- Created new abstract Export Action
- K4 export options added
