+++ proarc-webapp/src/main/java/cz/cas/lib/proarc/webapp/client/action/export/ArchiveExportAction.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package cz.cas.lib.proarc.webapp.client.action;
+package cz.cas.lib.proarc.webapp.client.action.export;
 
 import com.smartgwt.client.data.DSCallback;
 import com.smartgwt.client.data.DSRequest;
@@ -26,7 +26,8 @@ import com.smartgwt.client.types.PromptStyle;
 import com.smartgwt.client.util.SC;
 import cz.cas.lib.proarc.webapp.client.ClientMessages;
 import cz.cas.lib.proarc.webapp.client.ClientUtils;
-import cz.cas.lib.proarc.webapp.client.action.DesaExportAction.ExportResultWidget;
+import cz.cas.lib.proarc.webapp.client.action.ActionEvent;
+import cz.cas.lib.proarc.webapp.client.action.Actions;
 import cz.cas.lib.proarc.webapp.client.ds.DigitalObjectDataSource.DigitalObject;
 import cz.cas.lib.proarc.webapp.client.ds.ExportDataSource;
 import cz.cas.lib.proarc.webapp.client.ds.RestConfig;
@@ -37,17 +38,14 @@ import cz.cas.lib.proarc.webapp.shared.rest.ExportResourceApi;
  *
  * @author Jan Pokorsky
  */
-public class ArchiveExportAction extends AbstractAction {
-
-    private final ClientMessages i18n;
+public class ArchiveExportAction extends ExportAction {
 
     public ArchiveExportAction(ClientMessages i18n) {
         this(i18n, i18n.ArchiveExportAction_Title(), null, i18n.ArchiveExportAction_Hint());
     }
 
     public ArchiveExportAction(ClientMessages i18n, String title, String icon, String tooltip) {
-        super(title, icon, tooltip);
-        this.i18n = i18n;
+        super(i18n, title, icon, tooltip);
     }
 
     @Override
@@ -113,7 +111,7 @@ public class ArchiveExportAction extends AbstractAction {
         dsRequest.setPromptStyle(PromptStyle.DIALOG);
         dsRequest.setPrompt(i18n.KrameriusExportAction_Add_Msg());
         DataSource ds = ExportDataSource.getArchive();
-        ds.addData(export, new DSCallback() {
+        dsAddData(ds, export, new DSCallback() {
 
             @Override
             public void execute(DSResponse response, Object rawData, DSRequest request) {
@@ -127,7 +125,7 @@ public class ArchiveExportAction extends AbstractAction {
                         }
                         SC.say(i18n.ExportResultWidget_Window_Title(), i18n.ArchiveExportAction_ExportDone_Msg(target));
                     } else {
-                        ExportResultWidget.showErrors(erl.toArray());
+                        DesaExportAction.ExportResultWidget.showErrors(erl.toArray());
                     }
                 }
             }
