+++ proarc-webapp/src/main/java/cz/cas/lib/proarc/webapp/client/presenter/DigitalObjectManager.java
@@ -38,20 +38,20 @@ import cz.cas.lib.proarc.webapp.client.action.AbstractAction;
 import cz.cas.lib.proarc.webapp.client.action.ActionEvent;
 import cz.cas.lib.proarc.webapp.client.action.Actions;
 import cz.cas.lib.proarc.webapp.client.action.Actions.ActionSource;
-import cz.cas.lib.proarc.webapp.client.action.ArchiveExportAction;
-import cz.cas.lib.proarc.webapp.client.action.CejshExportAction;
-import cz.cas.lib.proarc.webapp.client.action.CrossrefExportAction;
-import cz.cas.lib.proarc.webapp.client.action.DataStreamExportAction;
 import cz.cas.lib.proarc.webapp.client.action.DeleteAction;
-import cz.cas.lib.proarc.webapp.client.action.DesaExportAction;
 import cz.cas.lib.proarc.webapp.client.action.DigitalObjectEditAction;
 import cz.cas.lib.proarc.webapp.client.action.FoxmlViewAction;
-import cz.cas.lib.proarc.webapp.client.action.KrameriusExportAction;
-import cz.cas.lib.proarc.webapp.client.action.NdkExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.NdkExportAction;
 import cz.cas.lib.proarc.webapp.client.action.RefreshAction;
 import cz.cas.lib.proarc.webapp.client.action.RefreshAction.Refreshable;
 import cz.cas.lib.proarc.webapp.client.action.TreeExpandAction;
 import cz.cas.lib.proarc.webapp.client.action.UrnNbnAction;
+import cz.cas.lib.proarc.webapp.client.action.export.ArchiveExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.CejshExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.CrossrefExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.DataStreamExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.DesaExportAction;
+import cz.cas.lib.proarc.webapp.client.action.export.KrameriusExportAction;
 import cz.cas.lib.proarc.webapp.client.ds.DigitalObjectDataSource;
 import cz.cas.lib.proarc.webapp.client.ds.MetaModelDataSource;
 import cz.cas.lib.proarc.webapp.client.ds.RelationDataSource;
