+++ proarc-webapp/src/main/java/cz/cas/lib/proarc/webapp/shared/rest/ExportResourceApi.java
@@ -38,6 +38,7 @@ public final class ExportResourceApi {
     public static final String KRAMERIUS4_PATH = "kramerius4";
     public static final String KRAMERIUS4_PID_PARAM = "pid";
     public static final String KRAMERIUS4_HIERARCHY_PARAM = "hierarchy";
+    public static final String KRAMERIUS4_POLICY_PARAM = "policy";
 
     public static final String DATASTREAM_PATH = "datastream";
     public static final String DATASTREAM_PID_PARAM = "pid";
