+++ proarc-webapp/src/main/resources/cz/cas/lib/proarc/webapp/client/ClientMessages.properties
@@ -669,7 +669,7 @@ ExportResultWidget_Window_Title=Validation Results
 ExportResultWidget_PID_Title=PID
 ExportResultWidget_Message_Title=Error Message
 
-KrameriusExportAction_Title=Kramerius 4 Export
+KrameriusExportAction_Title=Kramerius 4 Export...
 KrameriusExportAction_Hint=Exports selected digital objects in Kramerius 4 format.
 KrameriusExportAction_Add_Msg=Exporting. Please wait...
 ## {0} - target folder path
@@ -678,6 +678,12 @@ KrameriusExportAction_AddDone_Msg=Exported to user folder:<p/><i>{0}</i>
 NdkExportAction_Title=NDK PSP Export
 NdkExportAction_Hint=Exports selected objects in NDK PSP format.
 NdkExportAction_ExportDone_Msg=Exported successfully.
+
+ExportAction_Request_Title=Export settings
+ExportAction_Request_Workflow_Msg=Export entire workflow job
+ExportAction_Request_K4_Policy_Msg=Kramerius policy:
+ExportAction_Request_K4_Policy_Public=public
+ExportAction_Request_K4_Policy_Private=private
 
 RefreshAction_Title=Refresh
 RefreshAction_Hint=Reload data.
