+++ proarc-webapp/src/main/resources/cz/cas/lib/proarc/webapp/client/ClientMessages_cs.properties
@@ -670,7 +670,7 @@ ExportResultWidget_Window_Title=V\u00fdsledky validace
 ExportResultWidget_PID_Title=PID
 ExportResultWidget_Message_Title=Popis chyby
 
-KrameriusExportAction_Title=Export pro Kramerius 4
+KrameriusExportAction_Title=Export pro Kramerius 4...
 KrameriusExportAction_Hint=Exportuje vybran\u00e9 objekty ve form\u00e1tu Kramerius 4.
 KrameriusExportAction_Add_Msg=Prob\u00edh\u00e1 export. Pros\u00edm \u010dekejte...
 ## {0} - target folder path
@@ -679,6 +679,12 @@ KrameriusExportAction_AddDone_Msg=Exportov\u00e1no do u\u017eivatelsk\u00e9ho ad
 NdkExportAction_Title=NDK PSP Export
 NdkExportAction_Hint=Exportuje vybran\u00e9 objekty ve form\u00e1tu NDK PSP.
 NdkExportAction_ExportDone_Msg=Exportov\u00e1no.
+
+ExportAction_Request_Title=Nastaven\u00ed exportu
+ExportAction_Request_Workflow_Msg=Exportovat cel\u00fd z\u00e1m\u011br
+ExportAction_Request_K4_Policy_Msg=Viditelnost:
+ExportAction_Request_K4_Policy_Public=ve\u0159ejn\u00e9
+ExportAction_Request_K4_Policy_Private=priv\u00e1tn\u00ed
 
 RefreshAction_Title=Obnovit
 RefreshAction_Hint=Na\u010d\u00edst data znovu.
